+++ src/main/java/com/example/thymeleaf/repository/LightBulbRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LightBulbRepository extends JpaRepository<LightBulb, Integer> {
+public interface LightBulbRepository extends JpaRepository<LightBulb, Long> {
 
 }
